requirements.txt
accelerate==1.10.1
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
annoy==1.17.3
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.0.1
black==25.9.0
bleach==6.2.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.2.0
coverage==7.10.6
cryptography==45.0.5
dataclasses-json==0.6.7
datasets==4.0.0
Deprecated==1.2.18
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
durationpy==0.9
ecdsa==0.19.1
email_validator==2.2.0
etelemetry==0.3.1
execnet==2.1.1
Faker==37.8.0
fastapi==0.115.12
fastapi-cli==0.0.12
fastapi-cloud-cli==0.2.0
fastembed==0.6.0
filelock==3.18.0
flake8==7.3.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.0
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
groq==0.31.1
grpcio==1.71.0
h11==0.14.0
h2==4.2.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.2
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==32.0.1
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.75
langchain-groq==0.3.8
langchain-ollama==0.3.6
langchain-openai==0.3.32
langchain-text-splitters==0.3.7
langdetect==1.0.9
langid==1.1.6
langgraph==0.6.8
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.75
langchain-groq==0.3.8
langchain-ollama==0.3.6
langchain-openai==0.3.32
langchain-text-splitters==0.3.7
langgraph==0.6.8
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.3.45
lark==1.2.2
llama_cpp_python==0.3.8
loguru==0.7.3
looseversion==1.3.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
mypy==1.18.2
mypy-extensions==1.0.0

nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
ollama==0.5.1
onnxruntime==1.22.1
openai==1.108.1
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.16
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pathlib==1.0.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.4.0
pluggy==1.6.0
portalocker==3.2.0
posthog==3.23.0
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.5
prov==2.1.1
psutil==7.1.0
psycopg2==2.9.10
puremagic==1.30
py-cpuinfo==9.0.0
py_rust_stemmers==0.1.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydot==4.0.1
pyflakes==3.4.0
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.26.3
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PyStemmer==2.2.0.3
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-benchmark==5.1.0
pytest-cov==7.0.0
pytest-mock==3.15.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.1.10
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
qdrant-client==1.15.0
rdflib==7.1.4
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.0.0
sentencepiece==0.2.0
sentry-sdk==2.38.0
setuptools==80.9.0
shellingham==1.5.4
simpleeval==1.0.3
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
SQLAlchemy==2.0.39
starlette==0.46.1
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traits==7.0.2
transformers==4.56.2
triton==3.3.1
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==1.26.16
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
win32_setctime==1.2.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
